fix(optimize): support bytes() added in hono v4.13
